pyproject.toml
[project]
name = "imagehandler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "open3d>=0.19.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
